input fetch working
