+++ components/create-character/CChar-Form-2.vue
@@ -1,57 +1,115 @@
-// Appearance
+// Details
 
 <script setup lang="ts">
 import type { FormError, FormSubmitEvent } from "#ui/types";
 import { useAuthStore, useFormStore } from "#imports";
 
 type CharacterForm = {
-  hair: string;
-  skin: string;
-  eyes: string;
-  height: number;
-  weight: number;
-  age: number;
-  gender: string;
+  firstName: string;
+  lastName: string;
+  class: string;
+  alignment: string;
+  race: string;
 };
-const config = useRuntimeConfig();
 const authStore = useAuthStore();
 const formStore = useFormStore();
-const toast = useToast();
 const router = useRouter();
 
+const classes = [
+  { name: "Barbarian", value: "barbarian" },
+  { name: "Bard", value: "bard" },
+  { name: "Cleric", value: "cleric" },
+  { name: "Druid", value: "druid" },
+  { name: "Fighter", value: "fighter" },
+  { name: "Monk", value: "monk" },
+  { name: "Paladin", value: "paladin" },
+  { name: "Ranger", value: "ranger" },
+  { name: "Rogue", value: "rogue" },
+  { name: "Sorcerer", value: "sorcerer" },
+  { name: "Warlock", value: "warlock" },
+  { name: "Wizard", value: "wizard" },
+];
+
+const alignments = [
+  { name: "Chaotic Evil", value: "chaotic-evil" },
+  { name: "Chaotic Good", value: "chaotic-good" },
+  { name: "Chaotic Neutral", value: "chaotic-neutral" },
+  { name: "Lawful Evil", value: "lawful-evil" },
+  { name: "Lawful Good", value: "lawful-good" },
+  { name: "Lawful Neutral", value: "lawful-neutral" },
+  { name: "Neutral", value: "neutral" },
+  { name: "Neutral Evil", value: "neutral-evil" },
+  { name: "Neutral Good", value: "neutral-good" },
+];
+
+const races = [
+  { name: "Dragonborn", value: "dragonborn" },
+  { name: "Halfling", value: "halfling" },
+  { name: "Half-Elf", value: "half-elf" },
+  { name: "Half-Orc", value: "half-orc" },
+  { name: "Human", value: "human" },
+  { name: "Tiefling", value: "tiefling" },
+  { name: "---", value: "", disabled: true },
+  {
+    name: "Dwarves",
+    children: [
+      { name: "Dwarf", value: "dwarf" },
+      { name: "Hill Dwarf", value: "hill-dwarf" },
+      { name: "Mountain Dwarf", value: "mountain-dwarf" },
+    ],
+  },
+  {
+    name: "Elves",
+    children: [
+      { name: "Elf", value: "elf" },
+      { name: "Dark Elf (Drow)", value: "dark-elf" },
+      { name: "High Elf", value: "high-elf" },
+      { name: "Wood Elf", value: "wood-elf" },
+    ],
+  },
+  {
+    name: "Gnomes",
+    children: [
+      { name: "Gnome", value: "gnome" },
+      { name: "Forest Gnome", value: "forest-gnome" },
+      { name: "Rock Gnome", value: "rock-gnome" },
+    ],
+  },
+];
+
 const state = reactive({
-  hair: undefined,
-  skin: undefined,
-  eyes: undefined,
-  age: undefined,
-  height: undefined,
-  weight: undefined,
-  gender: undefined,
+  firstName: undefined,
+  lastName: undefined,
+  class: undefined,
+  alignment: undefined,
+  race: undefined,
+});
+
+const buttonText = reactive({
+  nextBtn: "Next Step: Appearance",
+  prevBtn: "Go Back",
 });
 
 const disabled = ref(false);
 const loading = ref(false);
-const buttonText = reactive({
-  nextBtn: "Create Character",
-  prevBtn: "Go Back",
-});
 
 const validate = (state: CharacterForm): FormError[] => {
   const errors = [];
-  if (state.hair && state.hair.length > 25)
-    errors.push({ path: "hair", message: "Must be less than 25 characters" });
-  if (state.skin && state.skin.length > 25)
-    errors.push({ path: "skin", message: "Must be less than 25 characters" });
-  if (state.eyes && state.eyes.length > 25)
-    errors.push({ path: "eyes", message: "Must be less than 25 characters" });
-  if (state.height && state.height < 1)
-    errors.push({ path: "height", message: "Must be at least 12 inches" });
-  if (state.gender && state.gender.length > 25)
-    errors.push({ path: "gender", message: "Must be less than 25 characters" });
-  if (state.age && state.age < 1)
-    errors.push({ path: "age", message: "Must be at least 1 year old" });
-  if (state.weight && state.weight < 1)
-    errors.push({ path: "weight", message: "Must be at least 1lb" });
+  if (!state.firstName) errors.push({ path: "firstName", message: "Required" });
+  if (state.firstName && state.firstName.length > 25)
+    errors.push({
+      path: "firstName",
+      message: "Must be less than 25 characters",
+    });
+  if (!state.lastName) errors.push({ path: "lastName", message: "Required" });
+  if (state.lastName && state.lastName.length > 25)
+    errors.push({
+      path: "lastName",
+      message: "Must be less than 25 characters",
+    });
+  if (!state.class) errors.push({ path: "class", message: "Required" });
+  if (!state.alignment) errors.push({ path: "alignment", message: "Required" });
+  if (!state.race) errors.push({ path: "race", message: "Required" });
   return errors;
 };
 
@@ -67,89 +125,71 @@ async function goBack() {
 }
 
 async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
-  const uuid = "506bdf8b-4906-48ea-ad18-4e65a8a02e59";
   const eventData = {
-    hair: event.data.hair,
-    skin: event.data.skin,
-    eyes: event.data.eyes,
-    age: event.data.age,
-    height: event.data.height,
-    weight: event.data.weight,
-    gender: event.data.gender,
+    firstName: event.data.firstName,
+    lastName: event.data.lastName,
+    class: event.data.class,
+    alignment: event.data.alignment,
+    race: event.data.race,
   };
   disabled.value = true;
   loading.value = true;
   await authStore.ensureValidToken();
   updateFormStore(eventData);
-  try {
-    const response = await $fetch<{
-      status: string;
-      message: string;
-      data: CharacterForm;
-    }>(config.public.baseURL + "/campaign/" + uuid + "/character/create", {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${authStore.token}`,
-        "Content-Type": "application/json",
-      },
-      body: formStore.formData,
-    });
-    console.log(response);
-    toast.add({
-      title: "Character Created!",
-      icon: "i-heroicons-check-circle-solid",
-    });
-    await navigateTo("/");
-  } catch (err) {
-    console.error(err);
-    toast.add({
-      title: "There was an error - please try again",
-      color: "red",
-      icon: "i-heroicons-x-circle-solid",
-    });
-  } finally {
-    loading.value = false;
-    disabled.value = false;
-    buttonText.nextBtn = "Create Character";
-    buttonText.prevBtn = "Back";
-    state.hair = undefined;
-    state.skin = undefined;
-    state.eyes = undefined;
-    state.age = undefined;
-    state.height = undefined;
-    state.weight = undefined;
-    state.gender = undefined;
-  }
+  await navigateTo("/cchar/cchar-3");
 }
 </script>
 
 <template>
   <UForm :validate="validate" :state="state" @submit.prevent="onSubmit">
-    <div class="grid grid-cols-2 gap-4">
-      <UFormGroup label="Hair" name="hair">
-        <UInput v-model="state.hair" :disabled="disabled" type="text" />
-      </UFormGroup>
-      <UFormGroup label="Skin" name="skin">
-        <UInput v-model="state.skin" :disabled="disabled" type="text" />
-      </UFormGroup>
-      <UFormGroup label="Eyes" name="eyes">
-        <UInput v-model="state.eyes" :disabled="disabled" type="text" />
-      </UFormGroup>
-      <UFormGroup label="Age (years)" name="age">
-        <UInput v-model="state.age" :disabled="disabled" type="number" />
-      </UFormGroup>
-      <UFormGroup label="Height (inches)" name="height">
-        <UInput v-model="state.height" :disabled="disabled" type="number" />
-      </UFormGroup>
-      <UFormGroup label="Weight (lbs)" name="weight" class="mb-4">
-        <UInput v-model="state.weight" :disabled="disabled" type="number" />
-      </UFormGroup>
-    </div>
-    <UFormGroup label="Gender" name="gender">
-      <UInput v-model="state.gender" :disabled="disabled" type="text" />
+    <UFormGroup label="Character First Name" name="firstName" class="mb-4">
+      <UInput
+        v-model="state.firstName"
+        placeholder="Must be less than 25 characters"
+        :disabled="disabled"
+        type="text"
+        required
+      />
     </UFormGroup>
-
-    <div class="flex flex-col justify-center">
+    <UFormGroup label="Character Last Name" name="lastName" class="mb-4">
+      <UInput
+        v-model="state.lastName"
+        placeholder="Must be less than 25 characters"
+        :disabled="disabled"
+        type="text"
+        required
+      />
+    </UFormGroup>
+    <UFormGroup label="Class" name="class" class="mb-4">
+      <USelect
+        v-model="state.class"
+        placeholder="Select a class..."
+        option-attribute="name"
+        :options="classes"
+        required
+      />
+    </UFormGroup>
+    <UFormGroup label="Race" name="race" class="mb-4">
+      <USelect
+        v-model="state.race"
+        placeholder="Select a race..."
+        :options="races"
+        :disabled="disabled"
+        option-attribute="name"
+        required
+      />
+    </UFormGroup>
+    <UFormGroup label="Alignment" name="alignment">
+      <USelect
+        v-model="state.alignment"
+        placeholder="Select an alignment..."
+        option-attribute="name"
+        :options="alignments"
+        :disabled="disabled"
+        required
+      />
+    </UFormGroup>
+    <div class="flex flex-col justify-center mb-4">
       <UButton
         block
         color="amber"
@@ -162,8 +202,7 @@ async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
       >
       <UButton
         type="submit"
-        block
-        class="p-2 box-border text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
+        class="p-2 box-border w-full text-white inline-flex h-[35px] items-center justify-center rounded-[4px] font-medium leading-none shadow-[0_2px_10px] focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none mt-[20px]"
         :loading="loading"
       >
         {{ buttonText.nextBtn }}</UButton
